added JSON repository and project refactored
