+++ CookieCookbook/App/IRecipesUserInteraction.cs
@@ -1,0 +1,13 @@
+﻿using CookieCookbook.Recipes;
+using CookieCookbook.Recipes.Ingredients;
+
+namespace CookieCookbook.App;
+
+public interface IRecipesUserInteraction
+{
+    void ShowMessage(string message);
+    void Exit();
+    void PrintExistingRecipes(IEnumerable<Recipe> allRecipes);
+    void PromptToCreateRecipe();
+    IEnumerable<Ingredient> ReadIngredientsFromUser();
+}
