+++ CookieCookbook/Data/FileFormat.cs
@@ -1,0 +1,7 @@
+﻿namespace CookieCookbook.Data;
+
+public enum FileFormat
+{
+    JSON,
+    TXT
+}
