+++ CookieCookbook/Recipes/Ingredients/IngredientsRegister.cs
@@ -1,0 +1,33 @@
+﻿namespace CookieCookbook.Recipes.Ingredients;
+
+public interface IIngredientsRegister
+{
+    IEnumerable<Ingredient> All { get; }
+
+    Ingredient? GetByID(int ingredientID);
+}
+
+public class IngredientsRegister : IIngredientsRegister
+{
+    public IEnumerable<Ingredient> All { get; } = new List<Ingredient>()
+    {
+        new WheatFlour(),
+        new CoconutFlour(),
+        new Butter(),
+        new Chocolate(),
+        new Sugar(),
+        new Cardamon(),
+        new Cinnamon(),
+        new CocoaPowder()
+    };
+
+    public Ingredient? GetByID(int ingredientID)
+    {
+        foreach (var ingredient in All)
+        {
+            if (ingredient.ID == ingredientID) return ingredient;
+        }
+
+        return null;
+    }
+}
