+++ CookieCookbook/Repositories/StringJsonRepository.cs
@@ -1,0 +1,12 @@
+﻿using CookieCookbook.Recipes;
+using System.Text.Json;
+
+namespace CookieCookbook.Repositories;
+
+public class StringJsonRepository : StringsRepository
+{
+    protected override List<string> TextToStrings(string fileContents) => JsonSerializer.Deserialize<List<string>>(fileContents);
+
+    protected override string StringsToText(List<string> strings) => JsonSerializer.Serialize(strings);
+
+}
